Refactored everything so that the matrix has doubles as entries. My sanity hangs by a thread.

+++ Matrix.h
@@ -10,32 +10,32 @@ public:
     // initialize a rows x cols matrix with all 0's
     Matrix(int rows, int cols);
     // initialize a rows x cols matrix according to a given 2D array or vector
-    Matrix(int rows, int cols, int** matrixArray);
+    Matrix(int rows, int cols, double** matrixArray);
     Matrix& operator=(const Matrix& src);
     Matrix(const Matrix& src);
     ~Matrix();
 
     // Operators
-    int operator()(int row, int col) const;
+    double operator()(int row, int col) const;
     Matrix& operator*(const Matrix& m) const;
 
     // Accessors
     int rows() const;
     int cols() const;
-    int valueAt(int row, int col) const;
-    int* rowVector(int row) const;
-    int* colVector(int col) const;
+    double valueAt(int row, int col) const;
+    double* rowVector(int row) const;
+    double* colVector(int col) const;
     bool isSquare() const;
     void print() const;
     Matrix& transpose() const;
 
     // Mutators
-    bool insert(int row, int col, int value);
+    bool insert(int row, int col, double value);
 
 private:
     int m_rows;
     int m_cols;
-    int** m_array;
+    double** m_array;
 };
 
 
@@ -55,7 +55,7 @@ bool Matrix::isSquare() const {
 }
 
 inline
-int Matrix::operator()(int row, int col) const {
+double Matrix::operator()(int row, int col) const {
     return this->valueAt(row, col);
 }
